requirements.txt
yfinance==0.2.65
python-telegram-bot==22.3
pandas==2.0.1
PyYAML==6.0.1
requests==2.32.4